requirements.txt
# Core Web Framework
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.12

# AI & LangChain
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
google-generativeai==0.8.5

# Vector Store & Embeddings
faiss-cpu==1.12.0
sentence-transformers==3.3.1

# Document Processing
pypdf==6.1.1

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
pydantic==2.10.6
pydantic-core==2.27.2